fix windows compile issues

+++ src/OpenLoco/src/World/Industry.cpp
@@ -318,7 +318,7 @@ namespace OpenLoco
             producedCargoMonthlyHistorySize[0]++;
         }
         const auto newValue = std::min<uint32_t>(producedCargoQuantityMonthlyTotal[0], 12750u) / 50;
-        producedCargoMonthlyHistory1[producedCargoMonthlyHistorySize[0] - 1] = newValue;
+        producedCargoMonthlyHistory1[producedCargoMonthlyHistorySize[0] - 1] = static_cast<uint8_t>(newValue);
 
         if (producedCargoMonthlyHistorySize[1] == std::size(producedCargoMonthlyHistory2))
         {
@@ -329,7 +329,7 @@ namespace OpenLoco
             producedCargoMonthlyHistorySize[1]++;
         }
         const auto newValue2 = std::min<uint32_t>(producedCargoQuantityMonthlyTotal[1], 12750u) / 50;
-        producedCargoMonthlyHistory2[producedCargoMonthlyHistorySize[1] - 1] = newValue2;
+        producedCargoMonthlyHistory2[producedCargoMonthlyHistorySize[1] - 1] = static_cast<uint8_t>(newValue2);
 
         producedCargoQuantityPreviousMonth[0] = producedCargoQuantityMonthlyTotal[0];
         producedCargoQuantityMonthlyTotal[0] = 0;
